Add progress output when deduplicating jobs
This introduces a simple logger interface so that callers can
optionally pass in a message handler and a progress meter.
Supporting the progress meter is the goal here and that is also
the reason we cannot simply use the default Python logging support.

Fixes https://github.com/thread/django-lightweight-queue/issues/40