Avoid redundant work in ReliableRedisBackend.startup
This aims to speed up queue startup when using this backend as
we have seen in production that this can take a while.

In particular if a master process is restarted while other machines
are still processing jobs then the startup on that machine can end
up retrying the move of in-flight jobs repeatedly while racing
against the valid processing by workers elsewhere.

This change limits the scope for this by inspecting the list of
processing queues and comparing it to the list of expected queues
in order to only rescue jobs from workers which no longer exist.