Modernise cron configuration import mechanics
As far as I can tell this is functionally equivalent to the previous
approach, however uses a mixture of supported Python and Django
import utils rather than relying on `__import__` (which is discouraged)
and the deprecated `imp` module.

This resolves deprecation warnings as well as some resource warnings
that originated from the previous `imp.find_module` call.