Reinstate typing by adding a settings protocol for local use
Introduce an long-name adapter which facades away the name munging
logic and allows re-use of this both for Django settings and custom
settings. This simplifies AppSettings which can now focus on just
layering the settings.

While this doesn't gain any type checking on the settings from the
layers, that was never present anyway.