Revert "Merge pull request #73 from thread/limit-keys-calls-on-startup"
This reverts commit 04b985f6c20d14c82324ebb51e733201b5fde91b, reversing
changes made to 6ff4b9083dfec9d80e2f7c6560770f40e65db625.

While using SCAN instead of KEYS is the recommended approach, it
turns out that it has worse real-world behaviour in the scenario
that the change was intended to make. Specifically the queue start-up
times were observed to be much longer while the load on the redis
instance was not reduced meaningfully, negatively impacting other
clients worse than when KEYS was used.